LilyPond data structure tests
generalizing the idea of the "gold standard" test, this test suite
transforms a subset of the existing visual tests into tests of equality
of textual data.

Visual tests are still needed to specify the expected output,
but the equality check between the hand-coded expected output and actual
lilygabc output can be evaluated in a fully automated fashion.