order of music property lists closer to LilyPond's
the code was built by mimicking output of \displayMusic,
but the actual property order resulting from (make-music ...)
invocation is inverse to the order of its arguments

One bug fix is included: CompletizeExtenderEvent was being built
incorrectly and had no effect.

For the first time there are some passing tests,
although only with the --lily-only option specified.