rename gold standard -> regression tests
the defining feature of "gold standard" tests is that structure and
contents of the standard are not understood and described, which is
never more true. We do understand the structures and describe them in
detail in an extensive test suite. The original "gold standard" test has
become a simple regression test.